+++ source/server/config_validation/api.cc
@@ -8,8 +8,8 @@ namespace Api {
 ValidationImpl::ValidationImpl(std::chrono::milliseconds file_flush_interval_msec)
     : Impl(file_flush_interval_msec) {}
 
-Event::DispatcherPtr ValidationImpl::allocateDispatcher() {
-  return Event::DispatcherPtr{new Event::ValidationDispatcher()};
+Event::DispatcherPtr ValidationImpl::allocateDispatcher(TimeSource& time_source) {
+  return Event::DispatcherPtr{new Event::ValidationDispatcher(time_source)};
 }
 
 } // namespace Api